requirements.txt
pandas == 1.1.2
numpy == v1.19.0
flask == 1.1.2
flask_restful==0.3.8
Flask-Migrate==2.7.0
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.12
python-dotenv==0.17.1
psycopg2-binary==2.8.6